backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lop-backend"
version = "0.1.0"
description = "FastAPI backend for the Life Observability Platform MVP."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic==1.16.5",
  "fastapi==0.118.0",
  "httpx==0.28.1",
  "psycopg[binary]==3.2.10",
  "pydantic-settings==2.11.0",
  "sqlalchemy==2.0.43",
  "uvicorn[standard]==0.37.0",
]

[project.optional-dependencies]
dev = [
  "pytest==8.4.2",
]

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
